graphql-api/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.9
asyncpg==0.27.0
pydantic==1.10.7
strawberry-graphql==0.130.0
email-validator==2.0.0